requirements.txt
openai>=1.40
python-dotenv>=1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0